// FightTeamControl/ClientApp/app/models/student.ts
export class Student {

    idStudent: number;
    name: string;
    surname: string;
    birthDate: Date;
    email: string;
    startDate: Date;
    hasHealthProblem: boolean;
    helthProblemDescription: string;
    useDrugs: boolean;
    drugsDescription: string;
    hasSurgery: boolean;
    surgeryDescription: string;
    bloodType: number;
    acceptedTerms: boolean;        
}